go go

+++ Sims/gfg.h
@@ -1,7 +1,7 @@
 
 #include<vector>
 
-std::vector<int> in_or_out(std::vector<float> refx,std::vector<float> refy, std::vector<float> newshx , std::vector<float> newshy) ;
+std::vector<int> in_or_out(std::vector<double> refx,std::vector<double> refy, std::vector<double> newshx , std::vector<double> newshy) ;
 
-float max(std::vector<float> yay);
-float min(std::vector<float> yay);
+double max(std::vector<double> yay);
+double min(std::vector<double> yay);
